requirements.txt
#-----------------------------------------------------------------
####### example-requirements.txt #######
#
###### Requirements without Version Specifiers ######
##-- nose
##-- nose-cov
##-- beautifulsoup4
#
###### Requirements with Version Specifiers ######
#   See https://www.python.org/dev/peps/pep-0440/#version-specifiers
##-- docopt == 0.6.1             # Version Matching. Must be version 0.6.1
##-- keyring >= 4.1.1            # Minimum version 4.1.1
##-- coverage != 3.5             # Version Exclusion. Anything except version 3.5
##-- Mopidy-Dirble ~= 1.1        # Compatible release. Same as >= 1.1, == 1.*
#
###### Refer to other requirements files ######
##-- -r other-requirements.txt
#
#
###### A particular file ######
##-- ./downloads/numpy-1.9.2-cp34-none-win32.whl
##-- http://wxpython.org/Phoenix/snapshot-builds/wxPython_Phoenix-3.0.3.dev1820+49a8884-cp34-none-win_amd64.whl
#
###### Additional Requirements without Version Specifiers ######
#   Same as 1st section, just here to show that you can put things in any order.
##-- rejected
##-- green
#--------------------------------------------------------------------





#Requirements for grovepi configuartion
cffi
smbus-cffi
RPi.GPIO
pyyaml

#from source
git+https://github.com/bivab/smbus-cffi.git
hg+http://hg.code.sf.net/p/raspberry-gpio-python/code#egg=RPi.GPIO
